Create MultiPlateMultiFileMergeLum20201216_special.py
Script for taking all of the lum files from the 20201216 run and compiling them to a single dataframe with barcodes and timestamps.  Note that the initial 2hr timepoint for WM88 is missing, therefore all the other lines have 30 reads but WM88 only has 29.  The script is modified from the general version for this specific case.